Refactor D4: share cancel handling via a CancellableEvent base
The six cancellable plugin events each repeated the cancelled field plus
isCancelled/setCancelled. Introduce shared/event/CancellableEvent holding
that boilerplate (subclasses still own their HandlerList as Bukkit
requires) and extend it from each event.
