+++ src/main/java/com/eternalcode/parcellockers/locker/event/LockerDeleteEvent.java
@@ -1,21 +1,19 @@
 package com.eternalcode.parcellockers.locker.event;
 
 import com.eternalcode.parcellockers.locker.Locker;
+import com.eternalcode.parcellockers.shared.event.CancellableEvent;
 import java.util.UUID;
-import org.bukkit.event.Cancellable;
-import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
 
 // Called when a locker is deleted. Fired synchronously on the main thread (see LockerManager#delete).
 // Warning: this event is not called when all lockers are deleted through "/parcel debug delete lockers" command
-public class LockerDeleteEvent extends Event implements Cancellable {
+public class LockerDeleteEvent extends CancellableEvent {
 
     private static final HandlerList HANDLER_LIST = new HandlerList();
 
     private final Locker locker;
     private final UUID player;
-    private boolean cancelled;
 
     public LockerDeleteEvent(Locker locker, UUID player) {
         this.locker = locker;
@@ -35,18 +33,7 @@ public class LockerDeleteEvent extends Event implements Cancellable {
     }
 
     @Override
-    public boolean isCancelled() {
-        return this.cancelled;
-    }
-
-    @Override
-    public void setCancelled(boolean cancel) {
-        this.cancelled = cancel;
-    }
-
-    @Override
     public @NotNull HandlerList getHandlers() {
         return HANDLER_LIST;
     }
-
 }
